+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## [Unreleased]
+### Added
+- Bundle Lerna with the manager
+
 ### Changed
 - Use `@absolunet/brand-guidelines` for documentation styles
 - Maintenance update